pyproject.toml
[project]
name = "genderclassificationmodel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "grad-cam>=1.5.5",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "opencv-python>=4.13.0.92",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
]
